moved recipe serializer tests to separate file, removed unused function
